Extract dll then execute alagent + refractor on AWS variable
