+++ ftalertlogicagent.csproj
@@ -42,6 +42,7 @@
     </Reference>
     <Reference Include="Newtonsoft.Json">
       <HintPath>..\packages\Newtonsoft.Json.6.0.8\lib\net45\Newtonsoft.Json.dll</HintPath>
+      <Package>monodevelop</Package>
     </Reference>
     <Reference Include="NodaTime">
       <HintPath>..\packages\NodaTime.1.3.0\lib\net35-Client\NodaTime.dll</HintPath>
@@ -55,9 +56,9 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="main.cs" />
     <Compile Include="GetDnsConfig.cs" />
-    <Compile Include="EmbeddedAssembly.cs" />
+    <Compile Include="alagent.cs" />
+    <Compile Include="main.py" />
   </ItemGroup>
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <ItemGroup>
@@ -72,6 +73,6 @@
     <EmbeddedResource Include="Resources\Newtonsoft.Json.dll" />
     <EmbeddedResource Include="Resources\Newtonsoft.Json.FSharp.dll" />
     <EmbeddedResource Include="Resources\NodaTime.dll" />
-    <EmbeddedResource Include="al_agent.msi" />
+    <EmbeddedResource Include="Resources\al_agent.msi" />
   </ItemGroup>
 </Project>
