+++ ftalertlogicagent.userprefs
@@ -1,13 +1,14 @@
 ﻿<Properties StartupItem="ftalertlogicagent.csproj">
   <MonoDevelop.Ide.Workspace ActiveConfiguration="Release|x86" />
-  <MonoDevelop.Ide.Workbench>
+  <MonoDevelop.Ide.Workbench ActiveDocument="alagent.cs">
     <Files>
-      <File FileName="main.cs" Line="7" Column="1" />
-      <File FileName="GetDnsConfig.cs" Line="2" Column="1" />
+      <File FileName="alagent.cs" Line="10" Column="38" />
     </Files>
   </MonoDevelop.Ide.Workbench>
   <MonoDevelop.Ide.DebuggingService.Breakpoints>
-    <BreakpointStore />
+    <BreakpointStore>
+      <Breakpoint file="/home/clyde/Documents/ftgit/ftalertlogicagent/ftalertlogicagent/ft-alertlogicagent-windows/main.py" line="1" column="1" />
+    </BreakpointStore>
   </MonoDevelop.Ide.DebuggingService.Breakpoints>
   <MonoDevelop.Ide.DebuggingService.PinnedWatches />
 </Properties>
